Write debug logs to a file (and move debug/assert code to extra header)
