+++ src/Qubic.vcxproj.filters
@@ -112,6 +112,9 @@
       <Filter>smart_contracts</Filter>
     </ClInclude>
     <ClInclude Include="tick_storage.h" />
+    <ClInclude Include="platform\debugging.h">
+      <Filter>platform</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <Filter Include="smart_contracts">
